[ENH,REF,ADD] check new transfer

+++ crontab/qc.sh
@@ -1,48 +1,70 @@
-#!/usr/bin/env bash
+!/usr/bin/env bash
 
 PATH=$PATH:/usr/bin
 
 ROOT_DIR=$1
 
+source $ROOT_DIR/crontab/functions.sh
+
+START_MESSAGE="Start QC script."
+print_current_time
+echo $START_MESSAGE
+
 mkdir -p $ROOT_DIR/outputs
 source $ROOT_DIR/crontab/env_qc_settings
 source $ROOT_DIR/crontab/.env_sta
-source $ROOT_DIR/crontab/timestamp_parsing.sh
 
 cd $ROOT_DIR
 
+# get timestamp latest execution QC
+
+QC_LOG=$ROOT_DIR/cron_out
+PATTERN_TIME_QC_LOG="\[\([0-9]\{4\}-[0-9]\{2\}-[0-9]\{2\} [0-9]\{2\}:[0-9]\{2\}:[0-9]\{2\}\)\].*"
+GREP_OUT_QC=$(grep_last_occurrences "$QC_LOG" "$START_MESSAGE" "1")
+TIME_PREVIOUS_QC=$(parse_date "$GREP_OUT_QC" "$PATTERN_TIME_QC_LOG" )
+
+# get timestamps file transfers
+
+DATA_TRANSFER_LOG=/home/RBINS.BE/bmdc/belgica_data_transfer/log.txt
+# TRANSFER_LOG=/home/nvds/Documents/RBINS/ODANext/qc_through_sensorthings/log_tmp.txt
+
+PATTERN_TIME_TRANSFER_LOG="\(^[0-9]\{4\}-[0-9]\{2\}-[0-9]\{2\} [0-9]\{2\}:[0-9]\{2\}:[0-9]\{2\}\).*"
+GREP_OUT_TRANSF="$(grep_last_occurrences "$TRANSFER_LOG" "QC " "10000")"
+
+PATTERN_TIME_QC_END=".*QC up to \([0-9]\{4\}-[0-9]\{2\}-[0-9]\{2\} [0-9]\{2\}:[0-9]\{2\}:[0-9]\{2\}\).*"
 DT_INT=$(echo $DT | tr -dc '0-9')
 DT_INT=$(($DT_INT))
 DT_UNIT=$(printf '%s\n' "${DT//[[:digit:]]/}")
 
-FMT="+%Y-%m-%d %H:%M:%S"
+COUNTER=0
+while read -r LINE; do
+    TIME_I=$(parse_date "$LINE" "$PATTERN_TIME_TRANSFER_LOG")
+    if [[ "$TIME_I" > "$TIME_PREVIOUS_QC" ]]; then
+        END_I=$(parse_date "$LINE" "$PATTERN_TIME_QC_END")
+        START_I=$(get_date "$END_I UTC -$((DT_INT+OVERLAP))$DT_UNIT")
 
-END_I=$(date -u --date="$END_I UTC" "$FMT")
-START_I=$(date -u --date="$END_I UTC -$((DT_INT+OVERLAP))$DT_UNIT" "$FMT")
+        CONTAINER_ID= $(docker run \
+                -d --rm --network=host --user "$(id -u):$(id -g)"\
+                --workdir /app \
+                -v "$CONFIG_FOLDER":/app/conf \
+                -v "$OUTPUT_FOLDER":/app/outputs \
+                -e DEV_SENSORS_USER="$SENSORS_USER" \
+                -e DEV_SENSORS_PASS="$SENSORS_PASS" \
+                rbinsbmdc/quality_assurance_tool:$IMAGE_TAG \
+                "time.start=$START_I" "time.end=$END_I")
+        
+        STATUS_CODE_CONTAINER="$(docker container wait $CONTAINER_ID)"
+        print_current_time
+        echo "Status code $CONTAINER_ID: $STATUS_CODE_CONTAINER"
 
-TIMESTAMP_NOW=$(date --date now "$FMT")
-echo -n "["$TIMESTAMP_NOW"] - "
-CONTAINER_ID= $(docker run \
-        -d --rm --network=host --user "$(id -u):$(id -g)"\
-        --workdir /app \
-        -v "$CONFIG_FOLDER":/app/conf \
-        -v "$OUTPUT_FOLDER":/app/outputs \
-        -e DEV_SENSORS_USER="$SENSORS_USER" \
-        -e DEV_SENSORS_PASS="$SENSORS_PASS" \
-        rbinsbmdc/quality_assurance_tool:$IMAGE_TAG \
-        "time.start=$START_I" "time.end=$END_I")
-
-STATUS_CODE_CONTAINER="$(docker container wait $CONTAINER_ID)"
-TIMESTAMP_NOW=$(date --date now "$FMT")
-echo -n "["$TIMESTAMP_NOW"] - "
-echo "Status code $CONTAINER_ID: $STATUS_CODE_CONTAINER"
-
-FMT_TRANSFER_SCRIPT="+%Y-%m-%d %H:%M:%S"
-export startdate=$(date -u --date="$START_I UTC" "$FMT_TRANSFER_SCRIPT")
-export enddate=$(date -u --date="$END_I UTC" "$FMT_TRANSFER_SCRIPT")
-
-TIMESTAMP_NOW=$(date --date now "$FMT")
-echo -n "["$TIMESTAMP_NOW"] - "
-echo "Start production "
-
-/usr/bin/env bash $ROOT_DIR/crontab/sta_raw_to_sta_prod_transfer\ 1.sh
+        FMT_TRANSFER_SCRIPT="+%Y-%m-%d %H:%M:%S"
+        startdate=$(date -u --date="$START_I UTC" "$FMT_TRANSFER_SCRIPT")
+        enddate=$(date -u --date="$END_I UTC" "$FMT_TRANSFER_SCRIPT")
+        
+        print_current_time
+        echo "Start production "
+        
+        sudo -i -u ndeville startdate="$startdate" enddate="$enddate"/usr/bin/env bash $ROOT_DIR/crontab/sta_raw_to_sta_prod_transfer\ 1.sh
+        COUNTER=$((COUNTER+1))
+    fi
+done <<< "$GREP_OUT_TRANSF"
